+++ changelog.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     ([PR](https://github.com/f-dangel/curvlinops/pull/149))
   - `ActivationHessianLinearOperator`
     ([PR](https://github.com/f-dangel/curvlinops/pull/157))
+- More test cases for `KFACInverseLinearOperator` ([PR](https://github.com/f-dangel/curvlinops/pull/158))
 
 ### Fixed/Removed
 
@@ -25,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   matrix-vector product) added to the eigenvalues corresponding to the bias parameters
   when `use_exact_damping=True` and `KFACLinearOperator._separate_weight_and_bias=True`
   ([PR](https://github.com/f-dangel/curvlinops/pull/156))
+- Bug in `KFACLinearOperator.load_state_dict` ([PR](https://github.com/f-dangel/curvlinops/pull/158))
 
 ### Internal